requirements.txt
requests
curl_cffi
fake_useragent
elevenlabs
selenium
numpy
pillow
moviepy
pytube
pydub